src/requirements.txt
Django==2.0.6
django-cors-headers==2.2.0
django-rest-swagger==2.1.2
djangorestframework
djangorestframework-jsonapi==2.4.0
djangorestframework-jwt==1.11.0
djoser==1.1.5
pytz==2018.5
psycopg2==2.7.3.2
